I think step 2 should be removed. Filter section.
On line 195 the sentence instructs us to "Modify your FilterButton <script> like this: "
```
<script>
   export let filter = 'all'
</script>
```

However, this is being received as a prop from out Todos component.  We should not be writing setting the variable 'filter' to anything in our FilterButton.svelte. We should just be exporting the variable 'filter' to indicate that we are receiving it as a prop, right ?